Refactor code to include a function to check if a URL is a YouTube link
The `tagger.py` script was updated to include a new function `is_yt()` that checks if a given URL is a YouTube link. This function uses regular expressions to match the URL pattern of YouTube videos. This change allows the code to conditionally append a timestamp to the URL when creating tags from a transcription, only if the URL is a YouTube link. This update enhances the flexibility and accuracy of URL timestamping in the tagging process.